Class representing XML source of a Carpentries episode

Wrapper around an xml document to manipulate and inspect Carpentries episodes

Details

This class is a fancy wrapper around the results of tinkr::to_xml() and has method specific to the Carpentries episodes.

Note

The current XLST spec for tinkr does not support kramdown, which the Carpentries Episodes are styled with, thus some block tags will be destructively modified in the conversion.

Public fields

path

[character] path to file on disk

yaml

[character] text block at head of file

body

[xml_document] an xml document of the episode

ns

[xml_document] an xml namespace set to the file name

Active bindings

show_problems

[list] a list of all the problems that occurred in parsing the episode

tags

[xml_nodeset] all the kramdown tags from the episode

questions

[character] the questions from the episode

keypoints

[character] the keypoints from the episode

objectives

[character] the objectives from the episode

challenges

[xml_nodeset] all the challenges blocks from the episode

solutions

[xml_nodeset] all the solutions blocks from the episode

output

[xml_nodeset] all the output blocks from the episode

error

[xml_nodeset] all the error blocks from the episode

code

[xml_nodeset] all the code blocks from the episode

name

[character] the name of the source file without the path

lesson

[character] the path to the lesson where the episode is from

Methods

Public methods

Episode$get_blocks()

Episode$label_divs()

Episode$get_divs()

Episode$get_yaml()

Episode$use_dovetail()

Episode$use_sandpaper()

Episode$remove_error()

Episode$remove_output()

Episode$move_objectives()

Episode$move_keypoints()

Episode$move_questions()

Episode$get_challenge_graph()

Episode$write()

Episode$reset()

Episode$isolate_blocks()

Episode$unblock()

Episode$new()

Episode$clone()

Method get_blocks()

return all block_quote elements within the Episode

Usage

Episode$get_blocks(type = NULL, level = 1L)

Arguments

type

the type of block quote in the Jekyll syntax like ".challenge", ".discussion", or ".solution"

level

the level of the block within the document. Defaults to 1, which represents all of the block_quotes are not nested within any other block quotes. Increase the nubmer to increase the level of nesting.

Returns

[xml_nodeset] all the blocks from the episode with the given tag and level.

Examples

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
# get all the challenges
scope$get_blocks(".challenge")
# get the solutions
scope$get_blocks(".solution", level = 2)
\dontrun{

  # download the source files for r-novice-gampinder into a Lesson object
  rng <- get_lesson("swcarpentry/r-novice-gapminder")
  dsp1 <- rng$episodes[["04-data-structures-part1.md"]]
  # There are 9 blocks in total
  dsp1$get_blocks()
  # One is a callout block
  dsp1$get_blocks(".callout")
  # One is a discussion block
  dsp1$get_blocks(".discussion")
  # Seven are Challenge blocks
  dsp1$get_blocks(".challenge")
  # There are eight solution blocks:
  dsp1$get_blocks(".solution", level = 2L)
}

Method label_divs()

label all the div elements within the Episode to extract them with $get_divs()

Usage

Episode$label_divs()

Method get_divs()

return all div elements within the Episode

Usage

Episode$get_divs(type = NULL)

Arguments

type

the type of div tag (e.g. 'challenge' or 'solution')

Method get_yaml()

Extract the yaml metadata from the episode

Usage

Episode$get_yaml()

Method use_dovetail()

Ammend or add a setup code block to use {dovetail}

This will convert your lesson to use the dovetail R package for processing specialized block quotes which will do two things:

convert your lesson from md to Rmd

add to your setup chunk the following code

library('dovetail')
source(dvt_opts())

If there is no setup chunk, one will be created. If there is a setup chunk, then the source and knitr_fig_path calls will be removed.

Usage

Episode$use_dovetail()

Method use_sandpaper()

Use the sandpaper package for processing

This will convert your lesson to use the {sandpaper} R package for processing the lesson instead of Jekyll (default). Doing this will have the following effects:

code blocks that were marked with liquid tags (e.g. {: .language-r} are converted to standard code blocks or Rmarkdown chunks (with language information at the top of the code block)

If rmarkdown is used and the lesson contains python code, library('reticulate') will be added to the setup chunk of the lesson.

Usage

Episode$use_sandpaper(rmd = FALSE)

Arguments

rmd

if TRUE, lessons will be converted to RMarkdown documents

Method remove_error()

Remove error blocks

Usage

Episode$remove_error()

Method remove_output()

Remove output blocks

Usage

Episode$remove_output()

Method move_objectives()

move the objectives yaml item to the body

Usage

Episode$move_objectives()

Method move_keypoints()

move the keypoints yaml item to the body

Usage

Episode$move_keypoints()

Method move_questions()

move the questions yaml item to the body

Usage

Episode$move_questions()

Method get_challenge_graph()

Create a graph of the top-level elements for the challenges.

Usage

Episode$get_challenge_graph(recurse = TRUE)

Arguments

recurse

if TRUE (default), the content of the solutions will be included in the graph; FALSE will keep the solutions as block_quote elements.

Returns

a data frame with four columns representing all the elements within the challenges in the Episode:

Block: The sequential number of the challenge block

from: the inward elements

to: the outward elements

pos: the position in the markdown document

Note that there are three special node names:

challenge: start or end of the challenge block

solution: start of the solution block

lesson: start of the lesson block

Examples

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$get_challenge_graph()

Method write()

write the episode to disk as markdown

Usage

Episode$write(path = NULL, format = "md", edit = FALSE)

Arguments

path

the path to write your file to. Defaults to an empty directory in your temporary folder

format

one of "md" (default) or "xml". This will create a file with the correct extension in the path

edit

if TRUE, the file will open in an editor. Defaults to FALSE.

Returns

the episode object

Examples

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$write()

Method reset()

Re-read episode from disk

Usage

Episode$reset()

Returns

the episode object

Examples

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
xml2::xml_text(scope$tags[1])
xml2::xml_set_text(scope$tags[1], "{: .code}")
xml2::xml_text(scope$tags[1])
scope$reset()
xml2::xml_text(scope$tags[1])

Method isolate_blocks()

Remove all elements except for those within block quotes that have a kramdown tag. Note that this is a destructive process.

Usage

Episode$isolate_blocks()

Returns

the Episode object, invisibly

Examples

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$body # a full document with block quotes and code blocks, etc
scope$isolate_blocks()$body # only one challenge block_quote

Method unblock()

convert challenge blocks to roxygen-like code blocks

Usage

Episode$unblock(token = "#'")

Arguments

token

the token to use to indicate non-code, Defaults to "#'"

Returns

the Episode object, invisibly

Examples

loop <- Episode$new(file.path(lesson_fragment(), "_episodes", "14-looping-data-sets.md"))
loop$body # a full document with block quotes and code blocks, etc
loop$get_blocks() # all the blocks in the episode
loop$unblock()
loop$get_blocks() # no blocks
loop$code # now there are two blocks with challenge tags

Method new()

Create a new Episode

Usage

Episode$new(path = NULL, process_tags = TRUE, fix_links = TRUE)

Arguments

path

[character] path to a markdown episode file on disk

process_tags

[logical] if TRUE (default), kramdown tags will be processed into attributes of the parent nodes. If FALSE, these tags will be treated as text

fix_links

[logical] if TRUE (default), links pointing to liquid tags (e.g. {{ page.root }}) and included links (those supplied by a call to {\% import links.md \%}) will be appropriately processed as valid links.

Returns

A new Episode object with extracted XML data

Examples

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$name
scope$lesson
scope$challenges

Method clone()

The objects of this class are cloneable with this method.

Usage

Episode$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Episode$get_blocks`
## ------------------------------------------------

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
# get all the challenges
scope$get_blocks(".challenge")

#> {xml_nodeset (2)}
#> [1] <block_quote sourcepos="45:1-60:14" ktag="{: .challenge}">\n  <heading so ...
#> [2] <block_quote sourcepos="62:1-95:14" ktag="{: .challenge}">\n  <heading so ...

# get the solutions
scope$get_blocks(".solution", level = 2)

#> {xml_nodeset (0)}

if (FALSE) {

  # download the source files for r-novice-gampinder into a Lesson object
  rng <- get_lesson("swcarpentry/r-novice-gapminder")
  dsp1 <- rng$episodes[["04-data-structures-part1.md"]]
  # There are 9 blocks in total
  dsp1$get_blocks()
  # One is a callout block
  dsp1$get_blocks(".callout")
  # One is a discussion block
  dsp1$get_blocks(".discussion")
  # Seven are Challenge blocks
  dsp1$get_blocks(".challenge")
  # There are eight solution blocks:
  dsp1$get_blocks(".solution", level = 2L)
}

## ------------------------------------------------
## Method `Episode$get_challenge_graph`
## ------------------------------------------------

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$get_challenge_graph()

#>   Block       from         to        pos level
#> 1     1  challenge    heading 45:1-60:14     1
#> 2     1    heading  paragraph 45:3-45:34     1
#> 3     1  paragraph code_block 47:3-48:68     1
#> 4     1 code_block     lesson  50:3-58:5     1
#> 5     2  challenge    heading 62:1-95:14     1
#> 6     2    heading  paragraph 62:3-62:27     1
#> 7     2  paragraph       list 64:3-64:55     1
#> 8     2       list code_block  66:3-72:1     1
#> 9     2 code_block     lesson  73:3-93:5     1

## ------------------------------------------------
## Method `Episode$write`
## ------------------------------------------------

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$write()

#> Creating temporary directory '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpNCbX4U/dir981371d919a'

## ------------------------------------------------
## Method `Episode$reset`
## ------------------------------------------------

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
xml2::xml_text(scope$tags[1])

#> [1] "{: .language-python}"

xml2::xml_set_text(scope$tags[1], "{: .code}")

#> {xml_nodeset (1)}
#> [1]  ktag="{: .language-python}"

xml2::xml_text(scope$tags[1])

#> [1] "{: .language-python}"

scope$reset()
xml2::xml_text(scope$tags[1])

#> [1] "{: .language-python}"

## ------------------------------------------------
## Method `Episode$isolate_blocks`
## ------------------------------------------------

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$body # a full document with block quotes and code blocks, etc

#> {xml_document}
#> <document sourcepos="1:1-95:14" xmlns="http://commonmark.org/xml/1.0">
#> [1] <heading sourcepos="1:1-1:81" level="2">\n  <text sourcepos="1:4-1:81" xm ...
#> [2] <list sourcepos="3:1-9:0" type="bullet" tight="true">\n  <item sourcepos= ...
#> [3] <code_block sourcepos="10:1-16:3" xml:space="preserve" name="" ktag="{: . ...
#> [4] <list sourcepos="19:1-27:0" type="bullet" tight="true">\n  <item sourcepo ...
#> [5] <code_block sourcepos="28:1-31:3" xml:space="preserve" name="" ktag="{: . ...
#> [6] <code_block sourcepos="33:1-35:3" xml:space="preserve" name="" ktag="{: . ...
#> [7] <code_block sourcepos="37:1-42:3" xml:space="preserve" name="" ktag="{: . ...
#> [8] <block_quote sourcepos="45:1-60:14" ktag="{: .challenge}">\n  <heading so ...
#> [9] <block_quote sourcepos="62:1-95:14" ktag="{: .challenge}">\n  <heading so ...

scope$isolate_blocks()$body # only one challenge block_quote

#> {xml_document}
#> <document sourcepos="1:1-95:14" xmlns="http://commonmark.org/xml/1.0">
#> [1] <block_quote sourcepos="45:1-60:14" ktag="{: .challenge}">\n  <heading so ...
#> [2] <block_quote sourcepos="62:1-95:14" ktag="{: .challenge}">\n  <heading so ...

## ------------------------------------------------
## Method `Episode$unblock`
## ------------------------------------------------

loop <- Episode$new(file.path(lesson_fragment(), "_episodes", "14-looping-data-sets.md"))
loop$body # a full document with block quotes and code blocks, etc

#> {xml_document}
#> <document sourcepos="1:1-182:38" xmlns="http://commonmark.org/xml/1.0">
#>  [1] <heading sourcepos="2:1-2:65" level="2">\n  <text sourcepos="2:4-2:9" xm ...
#>  [2] <list sourcepos="4:1-6:0" type="bullet" tight="true">\n  <item sourcepos ...
#>  [3] <code_block sourcepos="7:1-12:3" xml:space="preserve" name="" ktag="{: . ...
#>  [4] <code_block sourcepos="14:1-33:3" xml:space="preserve" name="" ktag="{:  ...
#>  [5] <heading sourcepos="36:1-36:126" level="2">\n  <text sourcepos="36:4-36: ...
#>  [6] <list sourcepos="38:1-47:0" type="bullet" tight="true">\n  <item sourcep ...
#>  [7] <code_block sourcepos="48:1-51:3" xml:space="preserve" name="" ktag="{:  ...
#>  [8] <code_block sourcepos="53:1-57:3" xml:space="preserve" name="" ktag="{:  ...
#>  [9] <code_block sourcepos="60:1-62:3" xml:space="preserve" name="" ktag="{:  ...
#> [10] <code_block sourcepos="64:1-66:3" xml:space="preserve" name="" ktag="{:  ...
#> [11] <heading sourcepos="69:1-69:52" level="2">\n  <text sourcepos="69:4-69:7 ...
#> [12] <list sourcepos="71:1-73:0" type="bullet" tight="true">\n  <item sourcep ...
#> [13] <code_block sourcepos="74:1-78:3" xml:space="preserve" name="" ktag="{:  ...
#> [14] <code_block sourcepos="80:1-87:3" xml:space="preserve" name="" ktag="{:  ...
#> [15] <list sourcepos="90:1-94:0" type="bullet" tight="true">\n  <item sourcep ...
#> [16] <block_quote sourcepos="95:1-108:14" ktag="{: .challenge}">\n  <heading  ...
#> [17] <block_quote sourcepos="110:1-140:14" ktag="{: .challenge}">\n  <heading ...
#> [18] <block_quote sourcepos="142:1-170:14" ktag="{: .challenge}">\n  <heading ...
#> [19] <heading sourcepos="172:1-172:29" level="3">\n  <text sourcepos="172:5-1 ...
#> [20] <html_block sourcepos="174:1-174:86" xml:space="preserve">&lt;img src="h ...
#> ...

loop$get_blocks() # all the blocks in the episode

#> {xml_nodeset (3)}
#> [1] <block_quote sourcepos="95:1-108:14" ktag="{: .challenge}">\n  <heading s ...
#> [2] <block_quote sourcepos="110:1-140:14" ktag="{: .challenge}">\n  <heading  ...
#> [3] <block_quote sourcepos="142:1-170:14" ktag="{: .challenge}">\n  <heading  ...

loop$unblock()
loop$get_blocks() # no blocks

#> {xml_nodeset (0)}

loop$code # now there are two blocks with challenge tags

#> {xml_nodeset (11)}
#>  [1] <code_block sourcepos="7:1-12:3" xml:space="preserve" name="" ktag="{: . ...
#>  [2] <code_block sourcepos="14:1-33:3" xml:space="preserve" name="" ktag="{:  ...
#>  [3] <code_block sourcepos="48:1-51:3" xml:space="preserve" name="" ktag="{:  ...
#>  [4] <code_block sourcepos="53:1-57:3" xml:space="preserve" name="" ktag="{:  ...
#>  [5] <code_block sourcepos="60:1-62:3" xml:space="preserve" name="" ktag="{:  ...
#>  [6] <code_block sourcepos="64:1-66:3" xml:space="preserve" name="" ktag="{:  ...
#>  [7] <code_block sourcepos="74:1-78:3" xml:space="preserve" name="" ktag="{:  ...
#>  [8] <code_block sourcepos="80:1-87:3" xml:space="preserve" name="" ktag="{:  ...
#>  [9] <code_block sourcepos="115:3-123:5" xml:space="preserve" name="" ktag="{ ...
#> [10] <code_block sourcepos="129:5-137:7" xml:space="preserve" name="" ktag="{ ...
#> [11] <code_block sourcepos="152:5-167:7" xml:space="preserve" name="" ktag="{ ...

## ------------------------------------------------
## Method `Episode$new`
## ------------------------------------------------

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$name

#> [1] "17-scope.md"

scope$lesson

#> [1] "/Users/runner/work/_temp/Library/pegboard/lesson-fragment"

scope$challenges

#> {xml_nodeset (2)}
#> [1] <block_quote sourcepos="45:1-60:14" ktag="{: .challenge}">\n  <heading so ...
#> [2] <block_quote sourcepos="62:1-95:14" ktag="{: .challenge}">\n  <heading so ...