Finds and fixes unresolved links within an Episode

LINKS: A list of XPath predicates

There are two reasons why commonmark would not recognise a link:

The link is a relative link and the anchor is in another ~~castle~~ file.

The link source uses a liquid variable that contains spaces.

Because of this, we need to find these unparsed links in the document and re-parse them as valid links.

fix_link_type(): Fixes all links in the document based on the type of link

find_lesson_link(): Finds all text nodes that contain the link type

resolve_links(): Operates on an individual text node within a paragraph.

modify underlying text to new nodes, splitting off links

inserts all the modified nodes before the text node

remove the old text node

links_within_text_regex(): finding different types of links within markdown text can be challenging because it involves characters used in regex for grouping and character classes. In general, I want to do two things with text that I get back from a document:

split the links out from the text

identify which parts of the resulting vector are links.

This way, I can convert the links to links and the text to text.

text_to_links(): Splits links away from text and returns a nodeset to insert

make_link(): makes a link depending on the link type

fix_links(body)

LINKS

fix_link_type(type, body)

find_lesson_links(body, type = "rel_link")

resolve_links(txt, type)

links_within_text_regex()

text_to_links(txt, ns = NULL, type, sourcepos = NULL)

make_link(txt, pattern, type = "rel_link")

Arguments

body
an XML document.

type
the type of link as defined in LINKS.

txt
text derived from xml2::xml_text()

ns
a namespace object

sourcepos
defaults to NULL. If this is not NULL, it's the sourcepos attribute of the text node(s) and will be applied to the new nodes.

pattern
a regular expression that is used for splitting the link from the surrounding text.

Format

An object of class list of length 4.

Value

fix_links(): the modified body

fix_link_type(): the xml document with new nodes added

find_lesson_links(): an xml nodeset of text nodes containing the links that match the link type

resolve_links(): modified text nodes

text_to_links(): if ns is NULL: a character vector of XML text nodes, otherwise, new XML text nodes.

Examples

helpers <- pegboard:::links_within_text_regex()
helpers

#>                                         to_split 
#> "(?<!(\\]|\\)|\\!))\\[|\\](?!(\\]|\\[|\\())|\\)" 
#>                                       find_links 
#>                           "(?<!\\[)\\](\\[|\\()" 

txt <- "text ![image text](a.png) with [a link](b.org) and text"
res <- strsplit(txt, helpers["to_split"], perl = TRUE)[[1]]
data.frame(res)

#>                        res
#> 1 text ![image text](a.png
#> 2                    with 
#> 3            a link](b.org
#> 4                 and text

grepl(helpers["find_links"], res, perl = TRUE)

#> [1]  TRUE FALSE  TRUE FALSE

txt <- "Some text [and a link]({{ page.root }}/link.to#thing), 
some other text [and another link][link-to-thing]."
pegboard:::text_to_links(txt, type = "rel_link")

#> [1] "<text>Some text </text>"                                                         
#> [2] "<link destination='{{ page.root }}/link.to#thing'><text>and a link</text></link>"
#> [3] "<text>, \nsome other text </text>"                                               
#> [4] "<link destination='link-to-thing'><text>and another link</text></link>"          
#> [5] "<text>.</text>"                                                                  

md <- c(md = "http://commonmark.org/xml/1.0")
class(md) <- "xml_namespace"
pegboard:::text_to_links(txt, md, "rel_link")

#> {xml_nodeset (5)}
#> [1] <text>Some text </text>
#> [2] <link destination="{{ page.root }}/link.to#thing">\n  <text>and a link</t ...
#> [3] <text>, \nsome other text </text>
#> [4] <link destination="link-to-thing">\n  <text>and another link</text>\n</link>
#> [5] <text>.</text>