Get a carpentries lesson in XML format

Download and extract a carpentries lesson in XML format. This uses git2r::clone() to download a carpentries lesson to your computer (defaults to the temporary directory and extracts the lesson in _episodes/ using tinkr::to_xml()

get_lesson(lesson = NULL, path = tempdir(), overwrite = FALSE, ...)

Arguments

lesson
a github user/repo pattern to point to the lesson

path
a directory to write the lesson to

overwrite
if the path exists, setting this to TRUE will overwrite the path, otherwise, the contents of the path will be returned if it is a lesson repository.

...
arguments passed on to Episode$new().

Value

a list of xml objects, one element per episode.

Examples

png <- get_lesson("swcarpentry/python-novice-gapminder")

#> cloning into '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpjEvVUV/swcarpentry--python-novice-gapminder'...
#> Receiving objects:   1% (83/8211),   50 kb
#> Receiving objects:  11% (904/8211),  347 kb
#> Receiving objects:  21% (1725/8211),  507 kb
#> Receiving objects:  31% (2546/8211), 3432 kb
#> Receiving objects:  41% (3367/8211), 3784 kb
#> Receiving objects:  51% (4188/8211), 4184 kb
#> Receiving objects:  61% (5009/8211), 4376 kb
#> Receiving objects:  71% (5830/8211), 4520 kb
#> Receiving objects:  81% (6651/8211), 4760 kb
#> Receiving objects:  91% (7473/8211), 6695 kb
#> Receiving objects: 100% (8211/8211), 6850 kb, done.

str(png, max.level = 1)

#> Classes 'Lesson', 'R6' <Lesson>
#>   Public:
#>     blocks: function (type = NULL, level = 0, path = FALSE) 
#>     challenges: function (path = FALSE, graph = FALSE, recurse = TRUE) 
#>     clone: function (deep = FALSE) 
#>     episodes: list
#>     files: active binding
#>     initialize: function (path = NULL, rmd = FALSE, ...) 
#>     isolate_blocks: function () 
#>     n_problems: active binding
#>     path: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpjEv ...
#>     reset: function () 
#>     rmd: FALSE
#>     show_problems: active binding
#>     solutions: function (path = FALSE) 
#>     thin: function (verbose = TRUE) 
#>   Private:
#>     deep_clone: function (name, value)