pegboard: Explore and Manipulate Markdown Curricula from the Carpentries

The Carpentries (https://carpentries.org) curricula is made of of lessons that are hosted as websites. Each lesson represents between a half day to two days of instruction and contains several episodes, which are written as 'kramdown'-flavored 'markdown' documents and converted to HTML using the 'Jekyll' static website generator. This package builds on top of the 'tinkr' package; reads in these markdown documents to 'XML' and stores them in R6 classes for convenient exploration and manipulation of sections within episodes.

See also

Author

Maintainer: Robert Davey robertdavey@carpentries.org (ORCID)

Authors:

Erin Becker erinbecker@carpentries.org (ORCID)

Zhian N. Kamvar zkamvar@carpentries.org (ORCID)

Other contributors:

Toby Hodges tobyhodges@carpentries.org [contributor]

Erin Becker ebecker@carpentries.org [contributor]

Kelly Barnes kbarnes@carpentries.org [contributor]

Robert Davey robertdavey@carpentries.org [contributor]