Class to contain a single Lesson by the Carpentries

This is a wrapper for several Episode class objects.

Details

Lessons are made of up several episodes within the _episodes/ directory of a lesson. This class keeps track of several episodes and allows us to switch between RMarkdown and markdown episodes

Public fields

path

[character] path to Lesson directory

episodes

[list] list of Episode class objects.

rmd

[logical] when TRUE, the episodes represent RMarkdown files, default is FALSE for markdown files.

Active bindings

n_problems

number of problems per episode

show_problems

contents of the problems per episode

files

the source files for each episode

Methods

Public methods

Lesson$blocks()

Lesson$challenges()

Lesson$solutions()

Lesson$thin()

Lesson$reset()

Lesson$isolate_blocks()

Lesson$new()

Lesson$clone()

Method blocks()

Gather all of the blocks from the lesson in a list of xml_nodeset objects

Usage

Lesson$blocks(type = NULL, level = 0, path = FALSE)

Arguments

type

the type of block quote in the Jekyll syntax like ".challenge", ".discussion", or ".solution"

level

the level of the block within the document. Defaults to 0, which represents all of the block_quotes within the document regardless of nesting level.

path

[logical] if TRUE, the names of each element will be equivalent to the path. The default is FALSE, which gives the name of each episode.

body

the XML body of a carpentries lesson (an xml2 object)

Method challenges()

Gather all of the challenges from the lesson in a list of xml_nodeset objects

Usage

Lesson$challenges(path = FALSE, graph = FALSE, recurse = TRUE)

Arguments

path

[logical] if TRUE, the names of each element will be equivalent to the path. The default is FALSE, which gives the name of each episode.

graph

[logical] if TRUE, the output will be a data frame representing the directed graph of elements within the challenges. See the get_challenge_graph() method in Episode.

recurse

[logical] when graph = TRUE, this will include the solutions in the output. See Episode for more details.

Method solutions()

Gather all of the solutions from the lesson in a list of xml_nodeset objects

Usage

Lesson$solutions(path = FALSE)

Arguments

path

[logical] if TRUE, the names of each element will be equivalent to the path. The default is FALSE, which gives the name of each episode.

Method thin()

Remove episodes that have no challenges

Usage

Lesson$thin(verbose = TRUE)

Arguments

verbose

[logical] if TRUE (default), the names of each episode removed is reported. Set to FALSE to remove this behavior.

Returns

the Lesson object, invisibly

Examples

frg <- Lesson$new(lesson_fragment())
frg$thin()

Method reset()

Re-read all Episodes from disk

Usage

Lesson$reset()

Returns

the Lesson object

Examples

frg <- Lesson$new(lesson_fragment())
frg$episodes[[1]]$body
frg$isolate_blocks()$episodes[[1]]$body # empty
frg$reset()$episodes[[1]]$body # reset

Method isolate_blocks()

Remove all elements except for those within block quotes that have a kramdown tag. Note that this is a destructive process.

Usage

Lesson$isolate_blocks()

Returns

the Episode object, invisibly

Examples

frg <- Lesson$new(lesson_fragment())
frg$isolate_blocks()$body # only one challenge block_quote

Method new()

create a new Lesson object from a directory

Usage

Lesson$new(path = NULL, rmd = FALSE, ...)

Arguments

path

[character] path to a lesson directory. This must have a folder called _episodes within that contains markdown episodes

rmd

[logical] when TRUE, the imported files will be the source RMarkdown files. Defaults to FALSE, which reads the rendered markdown files.

...

arguments passed on to Episode$new

Returns

a new Lesson object that contains a list of Episode objects in $episodes

Examples

frg <- Lesson$new(lesson_fragment())
frg$path
frg$episodes

Method clone()

The objects of this class are cloneable with this method.

Usage

Lesson$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Lesson$thin`
## ------------------------------------------------

frg <- Lesson$new(lesson_fragment())
frg$thin()

#> Removing 1 episode: 10-lunch.md

## ------------------------------------------------
## Method `Lesson$reset`
## ------------------------------------------------

frg <- Lesson$new(lesson_fragment())
frg$episodes[[1]]$body

#> {xml_document}
#> <document sourcepos="1:1-4:131" xmlns="http://commonmark.org/xml/1.0">
#> [1] <paragraph sourcepos="1:1-1:49">\n  <text sourcepos="1:1-1:49" xml:space= ...
#> [2] <list sourcepos="2:1-4:131" type="bullet" tight="true">\n  <item sourcepo ...

frg$isolate_blocks()$episodes[[1]]$body # empty

#> {xml_document}
#> <document sourcepos="1:1-4:131" xmlns="http://commonmark.org/xml/1.0">

frg$reset()$episodes[[1]]$body # reset

#> {xml_document}
#> <document sourcepos="1:1-4:131" xmlns="http://commonmark.org/xml/1.0">
#> [1] <paragraph sourcepos="1:1-1:49">\n  <text sourcepos="1:1-1:49" xml:space= ...
#> [2] <list sourcepos="2:1-4:131" type="bullet" tight="true">\n  <item sourcepo ...

## ------------------------------------------------
## Method `Lesson$isolate_blocks`
## ------------------------------------------------

frg <- Lesson$new(lesson_fragment())
frg$isolate_blocks()$body # only one challenge block_quote

#> NULL

## ------------------------------------------------
## Method `Lesson$new`
## ------------------------------------------------

frg <- Lesson$new(lesson_fragment())
frg$path

#> [1] "/Users/runner/work/_temp/Library/pegboard/lesson-fragment"

frg$episodes

#> $`10-lunch.md`
#> <Episode>
#>   Public:
#>     body: xml_document, xml_node
#>     challenges: active binding
#>     clone: function (deep = FALSE) 
#>     code: active binding
#>     error: active binding
#>     get_blocks: function (type = NULL, level = 1L) 
#>     get_challenge_graph: function (recurse = TRUE) 
#>     get_divs: function (type = NULL) 
#>     get_yaml: function () 
#>     initialize: function (path = NULL, process_tags = TRUE, fix_links = TRUE) 
#>     isolate_blocks: function () 
#>     lesson: active binding
#>     move_keypoints: function () 
#>     move_objectives: function () 
#>     move_questions: function () 
#>     name: active binding
#>     ns: http://commonmark.org/xml/1.0
#>     output: active binding
#>     path: /Users/runner/work/_temp/Library/pegboard/lesson-fragmen ...
#>     remove_error: function () 
#>     remove_output: function () 
#>     reset: function () 
#>     show_problems: active binding
#>     solutions: active binding
#>     tags: active binding
#>     unblock: function (token = "#'") 
#>     use_dovetail: function () 
#>     use_sandpaper: function (rmd = FALSE) 
#>     write: function (path = NULL, format = "md", edit = FALSE) 
#>     yaml: --- layout: break title: "Lunch" teaching: 0 exercises:  ...
#>   Private:
#>     clear_yaml_item: function (what) 
#>     deep_clone: function (name, value) 
#>     mutations: FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>     problems: list
#>     record_problem: function (x) 
#> 
#> $`12-for-loops.md`
#> <Episode>
#>   Public:
#>     body: xml_document, xml_node
#>     challenges: active binding
#>     clone: function (deep = FALSE) 
#>     code: active binding
#>     error: active binding
#>     get_blocks: function (type = NULL, level = 1L) 
#>     get_challenge_graph: function (recurse = TRUE) 
#>     get_divs: function (type = NULL) 
#>     get_yaml: function () 
#>     initialize: function (path = NULL, process_tags = TRUE, fix_links = TRUE) 
#>     isolate_blocks: function () 
#>     lesson: active binding
#>     move_keypoints: function () 
#>     move_objectives: function () 
#>     move_questions: function () 
#>     name: active binding
#>     ns: http://commonmark.org/xml/1.0
#>     output: active binding
#>     path: /Users/runner/work/_temp/Library/pegboard/lesson-fragmen ...
#>     remove_error: function () 
#>     remove_output: function () 
#>     reset: function () 
#>     show_problems: active binding
#>     solutions: active binding
#>     tags: active binding
#>     unblock: function (token = "#'") 
#>     use_dovetail: function () 
#>     use_sandpaper: function (rmd = FALSE) 
#>     write: function (path = NULL, format = "md", edit = FALSE) 
#>     yaml: --- title: "For Loops" teaching: 10 exercises: 15 questi ...
#>   Private:
#>     clear_yaml_item: function (what) 
#>     deep_clone: function (name, value) 
#>     mutations: FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>     problems: list
#>     record_problem: function (x) 
#> 
#> $`14-looping-data-sets.md`
#> <Episode>
#>   Public:
#>     body: xml_document, xml_node
#>     challenges: active binding
#>     clone: function (deep = FALSE) 
#>     code: active binding
#>     error: active binding
#>     get_blocks: function (type = NULL, level = 1L) 
#>     get_challenge_graph: function (recurse = TRUE) 
#>     get_divs: function (type = NULL) 
#>     get_yaml: function () 
#>     initialize: function (path = NULL, process_tags = TRUE, fix_links = TRUE) 
#>     isolate_blocks: function () 
#>     lesson: active binding
#>     move_keypoints: function () 
#>     move_objectives: function () 
#>     move_questions: function () 
#>     name: active binding
#>     ns: http://commonmark.org/xml/1.0 http://commonmark.org/xml/ ...
#>     output: active binding
#>     path: /Users/runner/work/_temp/Library/pegboard/lesson-fragmen ...
#>     remove_error: function () 
#>     remove_output: function () 
#>     reset: function () 
#>     show_problems: active binding
#>     solutions: active binding
#>     tags: active binding
#>     unblock: function (token = "#'") 
#>     use_dovetail: function () 
#>     use_sandpaper: function (rmd = FALSE) 
#>     write: function (path = NULL, format = "md", edit = FALSE) 
#>     yaml: --- title: "Looping Over Data Sets" teaching: 5 exercise ...
#>   Private:
#>     clear_yaml_item: function (what) 
#>     deep_clone: function (name, value) 
#>     mutations: FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>     problems: list
#>     record_problem: function (x) 
#> 
#> $`17-scope.md`
#> <Episode>
#>   Public:
#>     body: xml_document, xml_node
#>     challenges: active binding
#>     clone: function (deep = FALSE) 
#>     code: active binding
#>     error: active binding
#>     get_blocks: function (type = NULL, level = 1L) 
#>     get_challenge_graph: function (recurse = TRUE) 
#>     get_divs: function (type = NULL) 
#>     get_yaml: function () 
#>     initialize: function (path = NULL, process_tags = TRUE, fix_links = TRUE) 
#>     isolate_blocks: function () 
#>     lesson: active binding
#>     move_keypoints: function () 
#>     move_objectives: function () 
#>     move_questions: function () 
#>     name: active binding
#>     ns: http://commonmark.org/xml/1.0
#>     output: active binding
#>     path: /Users/runner/work/_temp/Library/pegboard/lesson-fragmen ...
#>     remove_error: function () 
#>     remove_output: function () 
#>     reset: function () 
#>     show_problems: active binding
#>     solutions: active binding
#>     tags: active binding
#>     unblock: function (token = "#'") 
#>     use_dovetail: function () 
#>     use_sandpaper: function (rmd = FALSE) 
#>     write: function (path = NULL, format = "md", edit = FALSE) 
#>     yaml: --- title: "Variable Scope" teaching: 10 exercises: 10 q ...
#>   Private:
#>     clear_yaml_item: function (what) 
#>     deep_clone: function (name, value) 
#>     mutations: FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>     problems: list
#>     record_problem: function (x) 
#>