Get a carpentries lesson in XML format

Download and extract a carpentries lesson in XML format. This uses gert::git_clone() to download a carpentries lesson to your computer (defaults to the temporary directory and extracts the lesson in _episodes/ using tinkr::to_xml()

get_lesson(lesson = NULL, path = tempdir(), overwrite = FALSE, ...)

Arguments

lesson
a github user/repo pattern to point to the lesson

path
a directory to write the lesson to

overwrite
if the path exists, setting this to TRUE will overwrite the path, otherwise, the contents of the path will be returned if it is a lesson repository.

...
arguments passed on to Episode$new().

Value

a list of xml objects, one element per episode.

Examples

png <- get_lesson("swcarpentry/python-novice-gapminder")

#> Loading required namespace: gert

#> Warning: incomplete final line found on '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp8bNyAz/swcarpentry--python-novice-gapminder/_episodes/04-built-in.md'

#> Warning: incomplete final line found on '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp8bNyAz/swcarpentry--python-novice-gapminder/_episodes/14-looping-data-sets.md'

str(png, max.level = 1)

#> Classes 'Lesson', 'R6' <Lesson>
#>   Public:
#>     blocks: function (type = NULL, level = 0, path = FALSE) 
#>     challenges: function (path = FALSE, graph = FALSE, recurse = TRUE) 
#>     clone: function (deep = FALSE) 
#>     episodes: list
#>     extra: NULL
#>     files: active binding
#>     initialize: function (path = NULL, rmd = FALSE, jekyll = TRUE, ...) 
#>     isolate_blocks: function () 
#>     n_problems: active binding
#>     path: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp8bN ...
#>     reset: function () 
#>     rmd: FALSE
#>     show_problems: active binding
#>     solutions: function (path = FALSE) 
#>     thin: function (verbose = TRUE) 
#>   Private:
#>     deep_clone: function (name, value)