Validate Links in a markdown document

This function will validate that links do not throw an error in markdown documents. This will include links to images and will respect robots.txt for websites.

validate_links(yrn, verbose = TRUE)

Arguments

yrn
a tinkr::yarn or Episode object.

verbose
if TRUE (default), messages will be printed as the validator works.

Value

a data frame containing the links, locations, and indicators if they passed tests

Details

External links

These links must start with a valid and secure protocol. This means that we will enforce HTTPS over HTTP. Any link with HTTP will be flagged. Most importantly, these links must not return an error code > 399.

Cross-lesson links

These links will have no protocol, but should resolve to the HTML version of a page and have the correct capitalisation.

Anchors (aka fragments)

Anchors are located at the end of URLs that start with a # sign. These are used to indicate a section of the documenation.

Alt-text (for images)

All images must have associated alt-text. In pandoc, this is acheived by writing the alt attribute in curly braces after the image: ![image caption](link){alt='alt text'}

Descriptive text

All links must have descriptive text associated with them, which is beneficial for screen readers scanning the links on a page to not have a list full of "link", "link", "link".